requirements.txt
albumentations @ git+https://github.com/albumentations-team/albumentations@aa53526fdb6c789f640635a1917d4e4dd47d1069
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
httplib2==0.14.0
idna==2.8
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.5.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyter-http-over-ws==0.0.8
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
netifaces==0.10.4
networkx==2.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
packaging==20.7
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2020.11.15
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.16.2
pycocotools==2.0.2
pycparser==2.20
pycups==1.9.73
Pygments==2.7.2
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.17.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.7.1
simplejson==3.16.0
six==1.15.0
ssh-import-id==5.10
systemd-python==234
terminado==0.9.1
testpath==0.4.4
tifffile==2020.11.18
toml==0.10.2
torch==1.7.0
torchvision==0.8.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.2.1
virtualenv-clone==0.5.4
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
